Accessibility fixes to pages

+++ VideoWeb/VideoWeb.AcceptanceTests/Hooks/DataSetupHooks.cs
@@ -40,12 +40,11 @@ namespace VideoWeb.AcceptanceTests.Hooks
 
             foreach (var conference in todaysConferences)
             {
-                if (!ClerkUserIsAParticipantInTheConference(conference.Participants, context.GetClerkUser().Username) ||
-                    !conference.Status.Equals(ConferenceState.Closed)) continue;
-
+                if (!ClerkUserIsAParticipantInTheConference(conference.Participants, context.GetClerkUser().Username)) continue;
+                
                 var hearingId = GetTheHearingIdFromTheConference(conference.Id, context);
 
-                if (HearingHasNotBeenDeletedAlready(hearingId, context))
+                if (HearingHasNotBeenDeletedAlready(hearingId, context) && !hearingId.Equals(Guid.Empty))
                     DeleteTheHearing(hearingId, context);
 
                 if (ConferenceHasNotBeenDeletedAlready(conference.Id, context))
@@ -68,7 +67,7 @@ namespace VideoWeb.AcceptanceTests.Hooks
             var conference = ApiRequestHelper.DeserialiseSnakeCaseJsonToResponse<ConferenceDetailsResponse>(context.Response.Content);
 
             if (conference.Hearing_id == null)
-                throw new DataMisalignedException("Hearing Id must be set");
+                return Guid.Empty;
 
             return (Guid)conference.Hearing_id;
         }
